% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bin2int.R
\name{bin2int}
\alias{bin2int}
\title{Convert binary mode to integers.}
\usage{
bin2int(bin)
}
\arguments{
\item{bin}{string vector in binary mode.}
}
\value{
A numeric vector.
}
\description{
Convert binary mode to integers.
}
\examples{
\dontrun{
bin2int("001101")
}
}
\author{
Jian Li <\email{rinds.book@gmail.com}>
}
\keyword{string}

