% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/biOps2Image.R
\name{biOps2Image}
\alias{biOps2Image}
\title{Convert imagedata object of biOps package to Image object of EBImage package.}
\usage{
biOps2Image(x)
}
\arguments{
\item{x}{imagedata object of biOps package.}
}
\value{
Image object.
}
\description{
Convert imagedata object to Image object.
}
\examples{
\dontrun{
biOps2Image(x)
}
}
\author{
Jian Li <\email{rinds.book@gmail.com}>
}

