\name{EarthquakeCN}
\alias{EarthquakeCN}
\docType{data}
\title{Earth quake data set}
\description{
Earth quake data set of China and its surrounding areas from 2014-09-08 to 2014-09-12.
}
\usage{data(EarthquakeCN)}
\format{
  A data frame with 6 columns.
  \describe{
    \item{\code{Time}}{Time of the earth quake.}
    \item{\code{Lat}}{Latitude of the earth quake.}
    \item{\code{Long}}{Longitude of the earth quake.}
    \item{\code{Dep}}{Depth of the earth quake.}
    \item{\code{Mag}}{Magnitude of the earth quake on the Richter scale.}
    \item{\code{Loc}}{Location of the earth quake.}
  }
}

\author{
  Jian Li <\email{rinds.book@gmail.com}>
}

\references{
  \url{http://data.earthquake.cn/datashare/globeEarthquake_csn.html}
}

\keyword{datasets}
