% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Image2biOps.R
\name{Image2biOps}
\alias{Image2biOps}
\title{Convert Image object of EBImage package to imagedata object of biOps package.}
\usage{
Image2biOps(x)
}
\arguments{
\item{x}{Image object of EBImage package.}
}
\value{
imagedata object.
}
\description{
Convert Image object to imagedata object.
}
\examples{
\dontrun{
Image2biOps(x)
}
}
\author{
Jian Li <\email{rinds.book@gmail.com}>
}

