% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/int2bin.R
\name{int2bin}
\alias{int2bin}
\title{Convert integers to binary mode.}
\usage{
int2bin(int, len = NULL)
}
\arguments{
\item{int}{integer vector.}

\item{len}{number of characters of the result string.}
}
\value{
A character vector.
}
\description{
Convert integers to binary mode.
}
\examples{
\dontrun{
int2bin(10)
}
}
\author{
Jian Li <\email{rinds.book@gmail.com}>
}
\keyword{string}

