% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotTS.R
\name{plotTS}
\alias{plotTS}
\title{Plot time series and its forecast values.}
\usage{
plotTS(series, forecast)
}
\arguments{
\item{series}{ts object.}

\item{forecast}{the output of predict or a forecast object.}
}
\value{
plot.
}
\description{
Plot time series and its forecast values.
}
\author{
Jian Li <\email{rinds.book@gmail.com}>
}

