% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sign.test.R
\name{sign.test}
\alias{sign.test}
\title{Sign test.}
\usage{
\method{sign}{test}(v, m0, h1 = c("two.sided", "less", "greater"),
  EXACT = c("exact", "approximation"))
}
\arguments{
\item{v}{numeric vector of data values. Non-finite (e.g. infinite or missing) values will be omitted.}

\item{m0}{a number specifying an optional parameter used to form the null hypothesis.}

\item{h1}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}

\item{EXACT}{whether to use exact method.}
}
\value{
A list with class "htest".
}
\description{
Sign test.
}
\examples{
\dontrun{
sign.test(rnorm(100))
}
}
\author{
Jian Li <\email{rinds.book@gmail.com}>
}

