\name{SogouCS08}
\alias{SogouCS08}
\docType{data}
\title{Sogou news data set}
\description{
News in 2008 from Sogou lab.
}
\usage{data(SogouCS08)}
\format{
  A data frame with 2 columns.
  \describe{
    \item{\code{Title}}{Title of news.}
    \item{\code{Content}}{Content of news.}
  }
}

\author{
  Jian Li <\email{rinds.book@gmail.com}>
}
\references{
  \url{http://www.sogou.com/labs/dl/cs.html}
}
\keyword{datasets}
